% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlotFromDBQuery.R
\name{agroMapVector}
\alias{agroMapVector}
\title{agroMapVector
The function agroMapVector visualises data on the grid of the Carpatclim reanalysis dataset with customised colorscale by using function trimColorSet.}
\usage{
agroMapVector(
  data,
  nticks = NULL,
  binwidth = NULL,
  minimum = NULL,
  maximum = NULL,
  roundPrecision = NULL,
  reverseColorScale = FALSE,
  colorSet = "RdYlGn",
  center = NULL,
  plotTitle = NULL,
  imageTitle = NULL,
  lonlat = FALSE
)
}
\arguments{
\item{data}{Numeric vector that contains data which will be plotted by agroMapVector function}

\item{nticks}{Number of colors in the colorscale (default setting is 6)}

\item{roundPrecision}{Number of decimal places to round off values on the colorbar}

\item{reverseColorScale}{If it is TRUE, the colorscale is reversed (default setting is FALSE)}

\item{center}{A number around which diverging colorscales will be centralised}

\item{plotTitle}{Main title can added to the plot}

\item{imageTitle}{File name of the plot, if NULL no graphics will be created.}

\item{lonlat}{Geographical latitudes and longitudes per 1° can be addded to the plot}

\item{colorset}{Name of the colorscale which is available in the list numBaseColors}

\item{minim}{Minium value of the visualised dataset}

\item{maxim}{Maximum value of the visualised dataset}
}
\value{

}
\description{
agroMapVector
The function agroMapVector visualises data on the grid of the Carpatclim reanalysis dataset with customised colorscale by using function trimColorSet.
}
\author{
Erzsebet Kristof, Roland Hollos
}
