% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromoBase.R
\name{agromoBaseUI}
\alias{agromoBaseUI}
\title{agromoBaseUI}
\usage{
agromoBaseUI(id)
}
\arguments{
\item{id}{of the GUI element}
}
\description{
The UI function for the BASE component of the GUI
}
\keyword{internal}
