% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteWindow.R
\name{agroMoSite}
\alias{agroMoSite}
\title{agroMoSite}
\usage{
agroMoSite(
  input,
  output,
  session,
  dataenv,
  baseDir,
  connection,
  centralData,
  languageState
)
}
\arguments{
\item{input}{environment which provides the results of the user input}

\item{output}{environment where the server output goes}

\item{session}{environment to get information about the current session}

\item{dataenv}{The central datastructure of the AgroMo}

\item{baseDir}{baseDir is the base directory for the modell inputs/outputs}
}
\description{
This function provides the server-logic for the SITE window
}
\keyword{internal}
