% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteWindow.R
\name{agroMoSiteUI}
\alias{agroMoSiteUI}
\title{agromositeui}
\usage{
agroMoSiteUI(id)
}
\arguments{
\item{id}{the id of the site window}
}
\description{
this function defines the user interface of the site window
}
\keyword{internal}
