% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{agroServer}
\alias{agroServer}
\title{agroServer}
\usage{
agroServer(input, output, session)
}
\arguments{
\item{input}{Environment of the GUI input elements}

\item{output}{Environment of the output code. This will we rendered}

\item{session}{environment which is a channel between R and the Web environment}
}
\description{
This is the server function for the agromo modell
}
\keyword{internal}
