% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteWindow.R
\name{agroTextInput}
\alias{agroTextInput}
\title{agroTextInut}
\usage{
agroTextInput(inputId, label, value = "", width = NULL, placeholder = NULL)
}
\description{
custom text input with html
}
\keyword{internal}
