% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAndPlot.R
\name{changingMGM}
\alias{changingMGM}
\title{changingMGM}
\usage{
changingMGM(mgmFile, baseDir, planting = NULL, harvest = NULL,
  fertilization = NULL, irrigation = NULL, grazing = NULL,
  mowing = NULL, thinning = NULL, planshift_date, planshift_density,
  harvshift_date, fertshift_date, irrshift_date, fertshift_amount,
  irrshift_amount)
}
\arguments{
\item{mgmFile}{asdf}
}
\description{
This function changes the management files
}
