% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDirStructure.R
\name{checkDirStucture}
\alias{checkDirStucture}
\title{checkDirStructure}
\usage{
checkDirStucture(chosenPath)
}
\arguments{
\item{chosenPath}{Filepath to validate for AgroMo main directory structure}
}
\description{
This function checks if the directory structure is correct for agromo
}
