% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromoGird.R
\name{decor_fun}
\alias{decor_fun}
\title{decor_fun}
\usage{
decor_fun(fun, tmpdir)
}
\arguments{
\item{fun}{the name of the function}

\item{tmpdir}{the temporary file for IPC}
}
\description{
This is a simple decorator for paralell progress bars
}
\keyword{internal}
