% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{displayProfile}
\alias{displayProfile}
\title{displayProfile}
\usage{
displayProfile(profName, startDate)
}
\arguments{
\item{profName}{the name of the curent profilePlot}
}
\description{
display profile graphs
}
\keyword{internal}
