% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromoGird.R
\name{fut_lapply}
\alias{fut_lapply}
\title{fut_lapply}
\usage{
fut_lapply(li, FUN, cores, ...)
}
\arguments{
\item{li}{starting list}

\item{FUN}{a decor_fun decorated function}

\item{cores}{number of cores}
}
\description{
parallel lapply alternative. the difference between this and future_parallel is that the result of this function is a future, not a value
}
\keyword{internal}
