% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilPlot.R
\name{getProfileVariables}
\alias{getProfileVariables}
\title{getProfileVariables}
\usage{
getProfileVariables(tag, centralData = getOption("AgroMo_centralData"))
}
\arguments{
\item{tag}{The profile tag}
}
\description{
Get profile variables of a given profiletag
}
\keyword{internal}
