% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertListToExcel.R
\name{listToExcel}
\alias{listToExcel}
\title{listToExcel}
\usage{
listToExcel(list_object)
}
\arguments{
\item{list_object}{The list to export}
}
\description{
This simple function exports a list to an excel document. The list elements will be worksheets in Excel
}
