% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiPlot}
\alias{multiPlot}
\title{multiPlot}
\usage{
multiPlot(
  input,
  output,
  session,
  measurement,
  outputNames,
  outputTable,
  experimentID,
  treatmentID,
  repetAvg = TRUE,
  connection,
  centralData,
  measAlias
)
}
\arguments{
\item{input}{sdfasdf}
}
\description{
multiPlot
}
