% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromoGird.R
\name{paraProgress}
\alias{paraProgress}
\title{paraProgress}
\usage{
paraProgress(
  X,
  FUN,
  cores,
  pb = txtProgressBar(min = 1, max = length(X), style = 3),
  pbUpdate = setTxtProgressBar,
  prepB = NULL,
  prepUpdate = NULL
)
}
\arguments{
\item{X}{starting list}

\item{FUN}{an arbitary 1D function}

\item{cores}{number of cores}

\item{pb}{Progress bar}

\item{pbUpdate}{Progress bar}
}
\description{
A simple function which creates the necessery files for the paraprogress bar
}
