% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromoGird.R
\name{paraWithProgress}
\alias{paraWithProgress}
\title{paraWithProgress}
\usage{
paraWithProgress(X, fun, cores, ...)
}
\arguments{
\item{X}{starting list}

\item{fun}{an arbitary 1D function}

\item{cores}{number of cores}
}
\description{
A simple function which creates the necessery files for the paraprogress bar
}
\keyword{internal}
