% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSingle.R
\name{plotSingle}
\alias{plotSingle}
\title{plotSingle}
\usage{
plotSingle(
  outputNames = NULL,
  dataenv,
  varName,
  timeFrame,
  groupFun,
  plotT = "scatter",
  conversionFactor = 1,
  measurement,
  experiment_id,
  treatment,
  repetationsAveraged,
  yTitle
)
}
\arguments{
\item{outputName}{blucs}
}
\description{
plotSingle
}
