% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromoGird.R
\name{readTable}
\alias{readTable}
\title{writeChainToDB}
\usage{
readTable(fName, variables, type, cell_id, numDays, startYear, endYear)
}
\arguments{
\item{fName}{the name of the file}

\item{variables}{The variable names}

\item{type}{.dayout or .annout}
}
\description{
This function reads the model binary and put that into a database
}
