% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAndPlot.R
\name{runAndPlot}
\alias{runAndPlot}
\title{runAndPlot}
\usage{
runAndPlot(
  input,
  output,
  session,
  baseDir,
  iniFile,
  weatherFile,
  soilFile,
  managementFile,
  outputName,
  planting,
  harvest,
  fertilization,
  irrigation,
  grazing,
  mowing,
  thinning,
  planshift_date,
  planshift_density,
  harvshift_date,
  fertshift_date,
  irrshift_date,
  fertshift_amount,
  irrshift_amount,
  connection,
  centralData,
  siteRun
)
}
\description{
Blabl
}
