% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAndPlot.R
\name{runAndPlotUI}
\alias{runAndPlotUI}
\title{runAndPlotUI}
\usage{
runAndPlotUI(id, label)
}
\arguments{
\item{id}{Unique identifier}

\item{label}{Label for the button}
}
\description{
This function provides the UI for the main runAndPlot function
}
