% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelConnection.R
\name{setupGUI}
\alias{setupGUI}
\title{setupGUI}
\usage{
setupGUI(iniName, baseDir, centralData)
}
\arguments{
\item{iniName}{The name of the inifile}

\item{baseDir}{The base directory from within the model will be launced}

\item{centralData}{The central datastructure of the AgroMo}
}
\description{
This function reads reneral setting from the iniFile
}
