% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAndPlot.R
\name{shiftFile}
\alias{shiftFile}
\title{shiftFile}
\usage{
shiftFile(manFile, destFile, shiftDate, varCol = NULL, shiftVar = NULL)
}
\arguments{
\item{manFile}{a}
}
\description{
Makes small changes: shiftes in management files
}
