% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizer.R
\name{stringSanitizer}
\alias{stringSanitizer}
\title{stringSanitizer}
\usage{
stringSanitizer(string)
}
\arguments{
\item{string}{string to sanitize}
}
\description{
This little function converts a given string to more sane format (no non-ASCII, no spacial, no whitespace)
}
\keyword{internal}
