% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlotFromDBQuery.R
\name{trimColorSet}
\alias{trimColorSet}
\title{trimColorSet
The function trimColorSet}
\usage{
trimColorSet(
  minim,
  maxim,
  center = NULL,
  nticks = 6,
  roundPrecision = NULL,
  reverseColorScale = FALSE,
  colorSet = "RdYlGn"
)
}
\arguments{
\item{minim}{Minium value of the visualised dataset}

\item{maxim}{Maximum value of the visualised dataset}

\item{center}{A number around which diverging colorscales will be centralised}

\item{nticks}{Number of colors in the colorscale (default setting is 6)}

\item{roundPrecision}{Number of decimal places to round off values on the colorbar}

\item{reverseColorScale}{If it is TRUE, the colorscale is reversed (default setting is FALSE)}

\item{colorset}{Name of the colorscale which is available in the list numBaseColors}
}
\value{

}
\description{
trimColorSet
The function trimColorSet
}
\author{
Roland Hollos, Erzsebet Kristof
}
