% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromoGird.R
\name{writeChainToDB}
\alias{writeChainToDB}
\title{writeChainToDB}
\usage{
writeChainToDB(
  baseDir,
  storyName,
  dbConnection,
  outputName,
  chainMatrix,
  variables,
  errorVector,
  type,
  climid
)
}
\arguments{
\item{dbConnection}{An SQLite connection}

\item{outputName}{The name of the result table}

\item{settings}{The result of the setupGUI}

\item{binaryName}{The name of the binary output file}
}
\description{
This function reads the model binary and put that into a database
}
