% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOutputs.R
\name{writeDataToDB}
\alias{writeDataToDB}
\title{writeDataToDB}
\usage{
writeDataToDB(settings, dbConnection, binaryName, outputName)
}
\arguments{
\item{settings}{The result of the setupGUI}

\item{dbConnection}{An SQLite connection}

\item{binaryName}{The name of the binary output file}

\item{outputName}{The name of the result table}
}
\description{
This function reads the model binary and put that into a database
}
