% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{getProfileVariables}
\alias{getProfileVariables}
\title{getProfileVariables}
\usage{
getProfileVariables(outputTable)

getProfileVariables(outputTable)
}
\arguments{
\item{outputTable}{is an outputTable}

\item{tag}{The profile tag}
}
\description{
Get profile variables of a given profiletag

Get profile variables of a given profiletag
}
\keyword{internal}
