% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlotFromDBQuery.R
\name{readQueryFromDB}
\alias{readQueryFromDB}
\title{readQueryFromDB}
\usage{
readQueryFromDB(connection, query, attachedDBS = NULL, queryModifiers = NULL)
}
\arguments{
\item{connection}{Database connection}

\item{query}{SQL query on the given database}

\item{attachedDBS}{Further databases which can be joined}
}
\description{
This function reads the database file with a given query, results a vector with 1104 elements
}
